% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect_theory.r
\name{prospect_theory}
\alias{prospect_theory}
\title{Prospect theory}
\usage{
prospect_theory(g1, g2)
}
\arguments{
\item{g1}{the first gamble in the form of a n x 2 matrix, where column 1 contains outcomes, columns 2 contains probabilities}

\item{g2}{the second gamble}
}
\value{
the probability to chose gambe 2
}
\description{
Implementation of prospect theory with parameters from Tversky & Kahneman (1992). 
The prospect_theory function calculates the choice probability
to select gamble 2 of two gambles.
The choice rule is a logit function.
}
